requirements.txt
rdflib
pyshacl
pdfminer.six
lightrag-hku
nano-vectordb
torch
transformers
huggingface_hub
